requirements.txt
python-telegram-bot>=21.0
apscheduler>=3.10
requests>=2.31
python-dotenv>=1.0
simmer-sdk>=0.9.0
eth-account>=0.13.0
py-order-utils>=0.3.0
py-clob-client>=0.34.0
